requirements.txt
alembic >= 1.0.0
disco-py >= 0.0.12
gevent == 1.2.2
python-dateutil >= 2.7.3
PyMySQL >= 0.9.2
SQLAlchemy >= 1.2.10
